requirements.txt
fastapi[all]
alembic
SQLAlchemy
psycopg2-binary
python-dotenv
sqlmodel
asyncpg
httpx
pytest
anyio
trio